apply changes only to component instead of wrapper
